Fix mobile nav bugs
Accidently deleted onscroll event for expanding the title bar. Set the 
scroll event to set border radius style, and have the open/close 
function to change it by adding a class. This way opening/closing the 
menu does not change the border radius when the title bar is at full 
width.